refactor doc ctor functions to doc utils. updated to lodash@4.0.0. Resolves #12
