+++ lib/couchdoc.js
@@ -1049,7 +1049,7 @@ CouchbaseDocument.$_findById = function (id, options, memo, fn) {
         return fn(err, [], misses);
       }
 
-      var modelObjs = _.map(results, self.$_createModelObject, self);
+      var modelObjs = _.map(results, _.bind(self.$_createModelObject, self));
       async.eachLimit(modelObjs, 10, function (modelObj, eachCB) {
         modelObj.$_populate(options, memo, function (err, popObj, missed) {
           if (missed && missed.length > 0) {
