+++ lib/document.js
@@ -1,13 +1,11 @@
 var inspect = require('util').inspect;
-var _ = require('lodash');
 var uuid = require('uuid');
 var EventEmitter = require('events').EventEmitter;
 
-var utils = require('./utils');
 var Schema = require('./schema');
 var Tree = require('./tree');
-var Type = require('./type');
-var cloneUtil = require('./clone-util');
+var utils = require('./utils');
+var docUtil = require('./doc-utils');
 
 module.exports = Document;
 
@@ -37,271 +35,17 @@ function Document(data) {
   // ensure the schema set
   if (!this.schema || !(this.schema instanceof Schema)) throw new TypeError('.schema hasn\'t been set');
 
-  var build = function (data, tree, object) {
-    tree = (tree instanceof Tree) ? tree : self.schema.tree;
-    object = (typeof object === 'object') ? object : this;
-    for (var prop in data) {
-      if (data.hasOwnProperty(prop)) {
-        // encapsulate each iteration in a scope
-        !function (prop) {
-          // if not in tree, return and continue on
-          if (!tree[prop]) return;
-          // if the property is an object, check if the tree property
-          // is a Tree instance object too
-          if (typeof data[prop] === 'object' && tree[prop] instanceof Tree) {
-            // define setter for object
-            utils.define(data[prop], 'set', {
-              writable: false,
-              enumerable: false,
-              configurable: false,
-              value: function (value) {
-                object[prop] = value;
-              }
-            });
-            // define getter for object
-            utils.define(data[prop], 'get', {
-              writable: false,
-              enumerable: false,
-              configurable: false,
-              value: function () {
-                return object[prop];
-              }
-            });
-
-            build(data[prop], tree[prop], object[prop]);
-          }
-          // we've reached some kind of scalar value
-          // that exists in the schema tree and the object
-          else {
-            object[prop] = data[prop];
-          }
-        }.call(this, prop);
-      }
-    }
-  }.bind(this);
+  var build = docUtil.build.bind(this);
 
   // overload refresh method on prototype
   var refresh = function () {
     if (utils.isFrozen(this)) return false;
-    var defineFromTree = function (tree, scope, table) {
-      var item;
-      for (item in tree) {
-        if (tree.hasOwnProperty(item)) {
-          !function (item) {
-            if (tree[item] === null || typeof tree[item] !== 'object') return;
-            // we don't want this as a possible field
-            if (tree[item].static) return;
-            if (tree[item].method) return;
+    var defineFromTree = docUtil.defineFromTree.bind(this);
 
-            if (tree[item].virtual === true) {
-              var getter = utils.isFunction(tree[item].value.get) ? tree[item].value.get : undefined;
-              var setter = utils.isFunction(tree[item].value.set) ? tree[item].value.set : undefined;
-
-              utils.define(scope, item, {
-                configurable: false,
-                enumerable: false,
-                get: getter,
-                set: setter
-              });
-
-              return;
-            }
-
-            // it must be an instance of Type
-            if (tree[item] instanceof Type) {
-              // only set on plain objects
-              if (!utils.isArray(scope)) {
-                // if it doesn't exist in the internal table
-                // then set it to undefined
-                table[item] = table[item] || undefined;
-                // create descriptor for property item on scope
-                // from tree descriptor
-                utils.define(scope, item, {
-                  configurable: false,
-                  enumerable: true,
-                  get: function () {
-                    return table[item] ? tree[item].get(table[item]) : undefined;
-                  },
-                  set: function (value) {
-                    if (utils.isFunction(tree[item].validate) && tree[item].validate(value)) {
-                      table[item] = tree[item].set(value);
-                      return table[item];
-                    } else {
-                      return false;
-                    }
-                  }
-                });
-              }
-
-              if (!utils.isUndefined(tree[item].default)) {
-                scope[item] = tree[item].default;
-              }
-            }
-
-            // if it is a tree instance then we need
-            // to do a recursive call to define the
-            // descriptors needed for the object
-            else if (tree[item] instanceof Tree || utils.isArray(tree[item])) {
-              table[item] = utils.isArray(tree[item]) ? [] : {};
-
-              if (utils.isArray(tree[item])) {
-                utils.define(scope, item, {
-                  configurable: false,
-                  enumerable: true,
-                  get: function () {
-                    return table[item];
-                  },
-                  set: function (value) {
-                    if (utils.isFunction(tree[item].validate) && tree[item].validate(value)) {
-                      table[item] = tree[item].set(value);
-                      return table[item];
-                    } else if (utils.isArray(value) && utils.isArray(tree[item]) && tree[item].type &&
-                      tree[item].type instanceof Type && utils.isFunction(tree[item].type.validate)) {
-                      var c = _.every(value, tree[item].type.validate, tree[item].type);
-                      if (c) {
-                        table[item] = value;
-                        return table[item];
-                      }
-                      else {
-                        return false;
-                      }
-                    } else if (Array.isArray(value) && _.isEmpty(tree[item])) {
-                      // just an array of object
-                      table[item] = value;
-                    } else {
-                      return false;
-                    }
-                  }
-                });
-
-                // custom set for the array so we can do obj.arrayProp.set([...]);
-                utils.define(scope[item], 'set', {
-                  configurable: false,
-                  enumerable: false,
-                  writable: false,
-                  value: function (value) {
-                    // use scope[item] so that the above set is invoked
-                    scope[item] = value;
-                    return scope[item];
-                  }
-                });
-              }
-              else {
-                // this doesn't allow sub objects to be set using assignment operator
-                // user.metadata = { foo: 'bar' };
-                // we have to do user.metadata.foo = 'bar';
-                // we can also use user.metadata.set({...});
-
-                utils.define(scope, item, {
-                  configurable: false,
-                  enumerable: true,
-                  writable: false,
-                  value: {}
-                });
-
-                // at least over ride inspect so we can properly get property
-                utils.define(scope[item], 'inspect', {
-                  configurable: false,
-                  enumerable: false,
-                  writable: false,
-                  value: function () {
-                    if (!Array.isArray(tree[item]) && !tree[item].Constructor) {
-                      var keys = Object.keys(tree[item]);
-                      var ret = {};
-                      keys.forEach(function (k) {
-                        ret[k] = table[item][k];
-                      });
-                      return ret;
-                    }
-                    else {
-                      return table[item];
-                    }
-                  }
-                });
-
-                // special set for nested objects
-                utils.define(scope[item], 'set', {
-                  configurable: false,
-                  enumerable: false,
-                  writable: false,
-                  value: function (value) {
-                    if (!tree[item].Constructor) {
-                      // it's just a plain object
-                      var allowedKeys = Object.keys(tree[item]);
-                      var dataToSet = _.pick(value, allowedKeys);
-                      allowedKeys.forEach(function (k) {
-                        if (tree[item][k].validate && utils.isFunction(tree[item][k].validate)) {
-                          if (tree[item][k].validate(dataToSet[k])) {
-                            table[item][k] = dataToSet[k];
-                          }
-                        }
-                      });
-
-                      return table[item];
-                    }
-                    else {
-                      table[item] = value;
-                      return table[item];
-                    }
-                  }
-                });
-              }
-
-              utils.define(scope[item], 'get', {
-                configurable: false,
-                enumerable: false,
-                writable: false,
-                value: function (key) {
-                  return table[item][key];
-                }
-              });
-
-              // overload array methods
-              if (utils.isArray(tree[item])) {
-                [
-                  'concat', 'every', 'filter', 'forEach', 'indexOf', 'join', 'lastIndexOf', 'map', 'pop',
-                  'push', 'reduce', 'reduceRight', 'reverse', 'shift', 'slice', 'some', 'sort', 'splice',
-                  'toString', 'unshift', 'valueOf'
-                ].map(function (method) {
-                    if (utils.isFunction([][method])) {
-                      utils.define(scope[item], method, {
-                        configurable: false,
-                        enumerable: false,
-                        writable: false,
-                        value: function (value) {
-                          if (value !== undefined) {
-                            if (tree[item].type instanceof Type && utils.isFunction(tree[item].type.validate)) {
-                              if (tree[item].type.validate(value)) {
-                                return [][method].apply(table[item], arguments);
-                              }
-                              else {
-                                return false;
-                              }
-                            }
-                            else {
-                              return [][method].apply(table[item], arguments);
-                            }
-                          }
-                          else {
-                            return [][method].apply(table[item], arguments);
-                          }
-                        }.bind(null)
-                      });
-                    }
-                  });
-              }
-
-              // recursive call to define descriptors
-              defineFromTree.call(self, tree[item], scope[item], table[item]);
-            }
-          }.call(self, item);
-        }
-      }
-    }.bind(this);
     // define
     defineFromTree(self.schema.tree, this, table);
     // free if in strict mode
-    if (this.schema.options.strict) utils.freeze(this);
+    if (self.schema.options.strict) utils.freeze(this);
   };
 
   // overload set method on prototype
@@ -367,7 +111,7 @@ function Document(data) {
   if ('object' == typeof data) {
     var prop;
     for (prop in data) {
-      if (utils.isArray(data[prop]) && utils.isArray(this[prop])) {
+      if (data.hasOwnProperty(prop) && utils.isArray(data[prop]) && utils.isArray(this[prop])) {
         data[prop].forEach(function (v) {
           this.push(v);
         }, this[prop]);
@@ -436,25 +180,23 @@ Document.prototype.$_toObject = function (options, json) {
 
   // When internally saving this document we always pass options,
   // bypassing the custom schema options.
-  if (!(options && 'Object' == utils.getFunctionName(options.constructor)) ||
+  if (!(options && 'Object' === utils.getFunctionName(options.constructor)) ||
     (options && options.$_useSchemaOptions)) {
     if (json) {
       options = this.schema.options.toJSON ?
-        cloneUtil.clone(this.schema.options.toJSON) :
-      {};
+        docUtil.clone(this.schema.options.toJSON) : {};
       options.json = true;
       options.$_useSchemaOptions = true;
     } else {
       options = this.schema.options.toObject ?
-        cloneUtil.clone(this.schema.options.toObject) :
-      {};
+        docUtil.clone(this.schema.options.toObject) : {};
       options.json = false;
       options.$_useSchemaOptions = true;
     }
   }
 
   for (var key in defaultOptions) {
-    if (options[key] === undefined) {
+    if (defaultOptions.hasOwnProperty(key) && options[key] === undefined) {
       options[key] = defaultOptions[key];
     }
   }
@@ -463,7 +205,7 @@ Document.prototype.$_toObject = function (options, json) {
   // to save it from being overwritten by sub-transform functions
   var originalTransform = options.transform;
 
-  var ret = cloneUtil.clone(this.$_table, options) || {};
+  var ret = docUtil.clone(this.$_table, options) || {};
 
   if (options.virtuals) {
     applyGetters(this, ret, 'virtuals', options);
@@ -527,7 +269,7 @@ function applyGetters(self, json, type, options) {
     for (var ii = 0; ii < plen; ++ii) {
       part = parts[ii];
       if (ii === last) {
-        branch[part] = cloneUtil.clone(self.get(path), options);
+        branch[part] = docUtil.clone(self.get(path), options);
       } else {
         branch = branch[part] || (branch[part] = {});
       }
