+++ package.json
@@ -30,7 +30,7 @@
     "debug": "^2.2.0",
     "hooks-fixed": "^1.1.0",
     "inflection": "^1.8.0",
-    "lodash": "^3.10.1",
+    "lodash": "^4.0.0",
     "memoizee":"~0.3.9",
     "mpath": "^0.2.1",
     "sliced": "^1.0.1",
